refactor: split endpoint spec service
